Refactored Helper Functions into /Resources
