+++ package.json
@@ -18,14 +18,14 @@
     "@cloudinary/url-gen": "^1.9.2",
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
-    "@prisma/client": "4.14.1",
+    "@prisma/client": "^4.15.0",
     "cloudinary": "^1.35.0",
     "eslint": "8.36.0",
     "eslint-config-next": "^13.2.4",
     "framer-motion": "^10.5.0",
     "next": "^13.4.4",
     "next-cloudinary": "^4.1.1",
-    "prisma": "^4.14.1",
+    "prisma": "^4.15.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.8.0"
