+++ prisma/schema.prisma
@@ -14,23 +14,25 @@ model User {
   id       String         @id @default(uuid())
   email    String         @unique
   password String
-  name     String?
+  name     String
   timers   Timer[] // One-to-many mapping
   guesses  CulpritGuess[]
   scores   Score?
-  hints    Hint?
+  hints    Hint[]
 }
 
 model Timer {
-  timerID   Int     @id @default(autoincrement())
-  user      User    @relation(fields: [userID], references: [id])
+  timerID   Int      @id @default(autoincrement())
+  user      User     @relation(fields: [userID], references: [id])
   userID    String
-  startTime Int     @default(0)
-  endTime   Int     @default(0)
-  totalTime Int     @default(0)
+  createdAt DateTime
+  startTime Int      @default(0)
+  endTime   Int      @default(0)
+  totalTime Int      @default(0)
   isActive  Boolean
 }
 
+// One-to-One
 model Score {
   scoreID      Int    @id @default(autoincrement())
   user         User   @relation(fields: [userID], references: [id])
@@ -42,16 +44,18 @@ model Score {
 }
 
 model CulpritGuess {
-  guessID     Int     @id @default(autoincrement())
-  user        User    @relation(fields: [userID], references: [id])
+  guessID     Int      @id @default(autoincrement())
+  user        User     @relation(fields: [userID], references: [id])
   userID      String
+  createdAt   DateTime
   culpritName String
   isCorrect   Boolean
 }
 
 model Hint {
-  hintID         Int    @id @default(autoincrement())
-  user           User   @relation(fields: [userID], references: [id])
-  userID         String @unique
-  hintsCollected Int    @default(0)
+  hintID      Int      @id @default(autoincrement())
+  user        User     @relation(fields: [userID], references: [id])
+  userID      String   @unique
+  collectedAt DateTime
+  hintName    String
 }
