+++ src/app/brandroom/page.jsx
@@ -3,7 +3,7 @@ import styles from "./components/styles.module.css";
 import { Container, Text, Box } from "@chakra-ui/react";
 import CldImage from "../components/CldImage";
 import { useEffect, useState } from "react";
-import fetchRoom from "@/pages/api/rooms/fetchRoom";
+import fetchRoom from "@/resources/fetchRoom";
 import Map from "../Map";
 
 export default function BrandRoom() {
