+++ src/app/clinic/page.jsx
@@ -6,7 +6,7 @@ import styles from "./components/styles.module.css";
 
 import background from "../../../public/Rooms/Clinic/clinic.png";
 import { useEffect, useState } from "react";
-import fetchRoom from "@/pages/api/rooms/fetchRoom";
+import fetchRoom from "@/resources/fetchRoom";
 import CldImage from "../components/CldImage";
 import Map from "../Map";
