--- src/pages/api/postChoices.js
@@ -1,15 +1,0 @@
-export default async function postChoices() {
-    const response = await fetch("/api/choices", {
-        method: "POST",
-        body: JSON.stringify({
-            // Input data fields 
-            // id,
-            // price,
-            // description,
-            // title,
-        }),
-    });
-    const data = response.json();
-
-    return data; 
-}
