+++ src/pages/api/prisma/user/index.js
@@ -1,10 +1,9 @@
 import prisma from "../../../../../lib/prisma";
 
 // POST /api/prisma/user
-// Required fields in body: email, password, timers, guesses
-// Optional fields in body: name, scores, hints
+// Required fields in body: email, password, name, timers, guesses
+// Optional fields in body: scores, hints
 export default async function handle(req, res) {
-  console.log(req.body);
 
   const result = await prisma.user.create({
     data: {
