+++ src/resources/updateTimer.js
